feat: user invoice crud

+++ entity/level.go
@@ -5,7 +5,7 @@ import "gorm.io/gorm"
 type Level struct {
 	ID              int     `json:"id" gorm:"primaryKey"`
 	Name            string  `json:"name"`
-	Discount        float32 `json:"discount"`
+	Discount        float64 `json:"discount"`
 	AvatarUrl       string  `json:"avatarUrl"`
 	MinTransanction int     `json:"minTransanction"`
 	gorm.Model      `json:"-"`
